多级评论的展示与实现

+++ src/main/java/com/seagold/community/entity/Question.java
@@ -11,7 +11,6 @@ package com.seagold.community.entity;
 
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
-import com.baomidou.mybatisplus.annotation.Version;
 import lombok.Data;
 
 /**
@@ -25,13 +24,13 @@ import lombok.Data;
 @Data
 public class Question {
     @TableId(type = IdType.AUTO)
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private int creator;
+    private Long creator;
     private int viewCount;
     private int commentCount;
     private int likeCount;
